brewmaster-java-domain/src/main/java/org/brewmaster/domain/Hop.java
package org.brewmaster.domain;

import net.sf.oval.constraint.NotEmpty;
import net.sf.oval.constraint.NotNull;

import javax.persistence.Entity;
import javax.persistence.Lob;
import javax.persistence.Table;
import javax.persistence.UniqueConstraint;

@Entity
@Table(uniqueConstraints = @UniqueConstraint(name = "UNIQUE_NAME", columnNames = "name"))
public class Hop extends BaseEntity {

	@NotNull(errorCode = "hop.description.null")
	@NotEmpty(errorCode = "hop.description.empty")
    @Lob
	private String description;

	public String getDescription() {
		return description;
	}

	public void setDescription(String description) {
		this.description = description;
	}

	
	@NotNull(errorCode = "hop.name.null")
	@NotEmpty(errorCode = "hop.name.empty")
	private String name;

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}
}
